refactor: remove old internal/pkg and adjust imports
